# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: optparts.kz
  version: 1.0.0
  description: |
    JWT-based authentication API with user management, product catalog, shopping cart, orders, and delivery addresses.

    ## Authentication
    Most endpoints require a Bearer JWT token. Obtain one via `/auth/login` or `/auth/register`, then pass it as:
    ```
    Authorization: Bearer <token>
    ```

    ## Roles
    | Role | Value |
    |------|-------|
    | User | `user` |
    | Manager | `manager` |
    | Admin | `admin` |
servers:
  - url: /api
    description: Local server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: c1d2e3f4-a5b6-7890-cdef-123456789012
        name:
          type: string
          example: Electronics
        description:
          type: string
          nullable: true
          example: Electronic devices and accessories
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ─── Brands ──────────────────────────────────────────────────────────────
    Brand:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: b1c2d3e4-f5a6-7890-bcde-123456789012
        name:
          type: string
          example: SoundMax
        description:
          type: string
          nullable: true
          example: Premium audio equipment
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ─── Products ────────────────────────────────────────────────────────────
    ProductImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filepath:
          type: string
          example: /uploads/products/1700000000000-abc.jpg
        order:
          type: integer
          example: 0
        productId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    ProductProperty:
      type: object
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
          example: Color
        value:
          type: string
          example: Midnight Black
        productId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: p1q2r3s4-t5u6-7890-vwxy-123456789012
        sku:
          type: string
          example: SKU-WH-001
        name:
          type: string
          example: Wireless Headphones
        price:
          type: number
          format: float
          example: 79.99
        categoryId:
          type: string
          format: uuid
          nullable: true
        category:
          $ref: '#/components/schemas/Category'
        brandId:
          type: string
          format: uuid
          nullable: true
        brand:
          $ref: '#/components/schemas/Brand'
        images:
          type: array
          items:
            $ref: '#/components/schemas/ProductImage'
        properties:
          type: array
          items:
            $ref: '#/components/schemas/ProductProperty'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ─── Cart ─────────────────────────────────────────────────────────────────
...
    Address:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: Home
        street:
          type: string
          example: 123 Abay Ave
        city:
          type: string
          example: Almaty
        postcode:
          type: string
          example: "050000"
        main:
          type: boolean
          example: true
        userId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ─── Orders ──────────────────────────────────────────────────────────────
    OrderItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
          nullable: true
        productName:
          type: string
          example: Wireless Headphones
        productSku:
          type: string
          example: SKU-WH-001
        priceAtOrder:
          type: number
          example: 79.99
          description: Price locked at the time the order was placed
        quantity:
          type: integer
          example: 2
        subtotal:
          type: number
          example: 159.98
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        address:
          $ref: '#/components/schemas/Address'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        status:
          type: string
          enum: [new, paid, pending, delivered, cancelled]
          example: new
        statusLabel:
          type: string
          example: Новый
          description: Human-readable Russian label for the status
        totalAmount:
          type: number
          example: 159.98
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ─── Errors ───────────────────────────────────────────────────────────────
    ValidationError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: array
          items:
            type: string
          example: ["email must be an email", "password must be at least 8 characters"]
        error:
          type: string
          example: Bad Request
    UnauthorizedError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 401
        message:
          type: string
          example: Access denied. Please log in.
        error:
          type: string
          example: Unauthorized
    ForbiddenError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 403
        message:
          type: string
          example: Insufficient permissions.
        error:
          type: string
          example: Forbidden
    NotFoundError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          type: string
          example: Resource not found.
        error:
          type: string
          example: Not Found
security:
  - bearerAuth: []
paths:
...
  /products:
    get:
      tags: [Products]
      summary: List all products
      security: []
      responses:
        '200':
          description: Array of products with images, properties, category and brand
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      tags: [Products]
      summary: Create product (admin / manager)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sku, name, price]
              properties:
                sku:
                  type: string
                  example: SKU-WH-001
                name:
                  type: string
                  example: Wireless Headphones
                price:
                  type: number
                  example: 79.99
                categoryId:
                  type: string
                  format: uuid
                  nullable: true
                brandId:
                  type: string
                  format: uuid
                  nullable: true
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
  /products/{id}:
    get:
      tags: [Products]
      summary: Get product by ID
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
    put:
      tags: [Products]
      summary: Update product (admin / manager)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number
                categoryId:
                  type: string
                  format: uuid
                  nullable: true
                brandId:
                  type: string
                  format: uuid
                  nullable: true
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
    delete:
      tags: [Products]
      summary: Delete product (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
  /products/{id}/images:
    post:
      tags: [Products]
      summary: Upload product images — saved to disk, filepath stored in DB (admin / manager)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [images]
              properties:
                images:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Images uploaded
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductImage'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
  /products/{id}/images/{imageId}:
    delete:
      tags: [Products]
      summary: Remove product image — deletes file from disk and DB row (admin / manager)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: imageId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Image deleted
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
  /products/{id}/properties:
    post:
      tags: [Products]
      summary: Add custom key-value property to product (admin / manager)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key, value]
              properties:
                key:
                  type: string
                  example: Color
                value:
                  type: string
                  example: Midnight Black
      responses:
        '201':
          description: Property added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductProperty'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
  /products/{id}/properties/{propId}:
    put:
      tags: [Products]
      summary: Update product property (admin / manager)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: propId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key, value]
              properties:
                key:
                  type: string
                  example: Color
                value:
                  type: string
                  example: Arctic White
      responses:
        '200':
          description: Property updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductProperty'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
    delete:
      tags: [Products]
      summary: Remove product property (admin / manager)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: propId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Property deleted
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'

  # ═══════════════════════════════════════════════════════════════════════════
  # CART
  # ═══════════════════════════════════════════════════════════════════════════
...
  /orders:
    get:
      tags: [Orders]
      summary: Get current user's orders
      responses:
        '200':
          description: Array of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
    post:
      tags: [Orders]
      summary: Place a new order — prices are locked at time of order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                addressId:
                  type: string
                  format: uuid
                  nullable: true
                  description: Must be one of the user's saved addresses
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [productId, quantity]
                    properties:
                      productId:
                        type: string
                        format: uuid
                      quantity:
                        type: integer
                        minimum: 1
                        example: 2
            example:
              addressId: a1b2c3d4-e5f6-7890-abcd-ef1234567890
              items:
                - productId: p1q2r3s4-t5u6-7890-vwxy-123456789012
                  quantity: 2
                - productId: b2c3d4e5-f6a7-8901-bcde-f12345678901
                  quantity: 1
      responses:
        '201':
          description: Order placed with locked-in prices
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Validation error or empty items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
  /orders/all:
    get:
      tags: [Orders]
      summary: List all orders — admin / manager only
      responses:
        '200':
          description: All orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get order by ID (own orders only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
  /orders/{id}/status:
    patch:
      tags: [Orders]
      summary: Update order status — admin / manager only
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [new, paid, pending, delivered, cancelled]
                  example: paid
      responses:
        '200':
          description: Order status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid status value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
  /orders/{id}/cancel:
    patch:
      tags: [Orders]
      summary: Cancel own order — not allowed if already delivered
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Cannot cancel a delivered order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'